+++ NEWS.md
@@ -1,3 +1,11 @@
+# treasury 0.2.0
+
+* Better docs
+* Support for Treasury High Quality Market (HQM) Corporate Bond Yield Curve data
+* Support for Treasury Nominal Coupon-Issue (TNC) Yield Curve data
+* Support for Treasury Real Coupon-Issue (TRC) Yield Curve data
+* Support for Treasury Breakeven Inflation Curve (TBI curve) data
+
 # treasury 0.1.0
 
 * Initial CRAN submission.